requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
altair==5.5.0
Appium-Python-Client==4.4.0
attrs==25.1.0
bcrypt==4.2.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
core-universal4==4.24.2
cycler==0.12.1
dnspython==2.6.1
email_validator==2.2.0
eyes-common==5.30.1
eyes-selenium==5.27.2
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.55.7
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.0
h11==0.14.0
idna==3.10
ImageHash==4.3.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyboard==0.13.5
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==2.21.0
matplotlib==3.10.0
mdurl==0.1.2
narwhals==1.24.1
numpy==2.2.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
pillow==11.1.0
plotly==6.0.0
pluggy==1.5.0
protobuf==5.29.3
pyarrow==19.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
pytest==8.3.3
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyWavelets==1.7.0
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scipy==1.14.1
seaborn==0.13.2
selenium==4.26.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.34
streamlit==1.41.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
WTForms==3.1.2
